# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fragrance Rater",
    "description": "HTTP API for the Fragrance Rater service: a personal fragrance evaluation and recommendation system. Provides the fragrance catalog and an LLM-powered rating endpoint that scores a fragrance description and returns the model's reasoning. The rating endpoint calls OpenRouter (default model: `anthropic/claude-3.5-sonnet`); set `TEST_MODE=true` to use the bundled fixture response instead, which is what the Newman contract tests rely on. `POST /ratings` requires a shared household `X-API-Key` header and is rate limited; read-only endpoints do not require it.",
    "contact": {
      "name": "Byron Williams",
      "url": "https://github.com/ByronWilliamsCPA/fragrance-rater",
      "email": "byron@williamshome.family"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recommendations/{reviewer_id}": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendations",
        "description": "Get personalized fragrance recommendations for a reviewer.\n\nRequires at least 3 evaluations to generate meaningful recommendations.\nReturns fragrances sorted by match score, with vetoed items last.",
        "operationId": "get_recommendations_api_v1_recommendations__reviewer_id__get",
        "parameters": [
          {
            "name": "reviewer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reviewer Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum recommendations",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum recommendations"
          },
          {
            "name": "exclude_rated",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Exclude already-rated fragrances",
              "default": true,
              "title": "Exclude Rated"
            },
            "description": "Exclude already-rated fragrances"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{reviewer_id}/profile": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Profile Summary",
        "description": "Get a summary of a reviewer's preference profile.\n\nShows top liked/disliked notes, preferred accords, and fragrance families\nbased on their evaluation history. Optionally includes an LLM-generated\nnatural language summary.\n\nRequires the shared household ``X-API-Key`` header (see\n``fragrance_rater.middleware.auth.require_api_key``) and is rate limited\n(see ``fragrance_rater.middleware.rate_limit.RATINGS_RATE_LIMIT``)\nbecause ``include_llm=True`` can trigger a billed OpenRouter call.\n\nArgs:\n    request (Request): Incoming request, required by the\n        ``@limiter.limit`` decorator to key rate limiting on the caller.\n    reviewer_id (str): Reviewer to summarize.\n    session (Annotated[AsyncSession, Depends(get_db)]): Database session, injected.\n    service (Annotated[RecommendationService, Depends(get_recommendation_service)]):\n        Recommendation service, injected.\n    llm_service (Annotated[LLMService, Depends(get_llm_service)]): LLM service, injected.\n    include_llm (Annotated[bool, Query(description='Include LLM-generated summary')]):\n        Whether to include an LLM-generated summary.\n\nReturns:\n    ProfileSummaryResponse: The reviewer's preference profile summary.",
        "operationId": "get_profile_summary_api_v1_recommendations__reviewer_id__profile_get",
        "parameters": [
          {
            "name": "reviewer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reviewer Id"
            }
          },
          {
            "name": "include_llm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include LLM-generated summary",
              "default": true,
              "title": "Include Llm"
            },
            "description": "Include LLM-generated summary"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-API-Key header. Required on this endpoint because it can trigger a billed LLM call (see fragrance_rater.middleware.auth)."
          },
          "429": {
            "description": "Rate limit exceeded; retry after a short delay."
          },
          "503": {
            "description": "No household API key is configured for this deployment (FRAGRANCE_RATER_API_KEY unset)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{reviewer_id}/{fragrance_id}/explain": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendation Explanation",
        "description": "Get an LLM-generated explanation for why a fragrance matches a reviewer.\n\nUses the reviewer's preference profile and the fragrance's notes/accords\nto generate a personalized explanation.\n\nRequires the shared household ``X-API-Key`` header (see\n``fragrance_rater.middleware.auth.require_api_key``) and is rate limited\n(see ``fragrance_rater.middleware.rate_limit.RATINGS_RATE_LIMIT``)\nbecause this endpoint always triggers a billed OpenRouter call.\n\nArgs:\n    request (Request): Incoming request, required by the\n        ``@limiter.limit`` decorator to key rate limiting on the caller.\n    reviewer_id (str): Reviewer whose preference profile is used.\n    fragrance_id (str): Fragrance to explain the match for.\n    session (Annotated[AsyncSession, Depends(get_db)]): Database session, injected.\n    service (Annotated[RecommendationService, Depends(get_recommendation_service)]):\n        Recommendation service, injected.\n    llm_service (Annotated[LLMService, Depends(get_llm_service)]): LLM service, injected.\n\nReturns:\n    ExplanationResponse: The LLM-authored explanation.\n\nRaises:\n    HTTPException: 400 if the reviewer has fewer than 3 evaluations, or\n        404 if `fragrance_id` does not resolve to a live (non-soft-\n        deleted) fragrance.",
        "operationId": "get_recommendation_explanation_api_v1_recommendations__reviewer_id___fragrance_id__explain_get",
        "parameters": [
          {
            "name": "reviewer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reviewer Id"
            }
          },
          {
            "name": "fragrance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fragrance Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-API-Key header. Required on this endpoint because it triggers a billed LLM call (see fragrance_rater.middleware.auth)."
          },
          "429": {
            "description": "Rate limit exceeded; retry after a short delay."
          },
          "503": {
            "description": "No household API key is configured for this deployment (FRAGRANCE_RATER_API_KEY unset)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExplanationResponse": {
        "properties": {
          "fragrance_id": {
            "type": "string",
            "title": "Fragrance Id"
          },
          "fragrance_name": {
            "type": "string",
            "title": "Fragrance Name"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model used to generate explanation"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "Whether response was cached",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "fragrance_id",
          "fragrance_name",
          "explanation",
          "model"
        ],
        "title": "ExplanationResponse",
        "description": "Response model for recommendation explanation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileSummaryResponse": {
        "properties": {
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "evaluation_count": {
            "type": "integer",
            "title": "Evaluation Count"
          },
          "top_liked_notes": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Top Liked Notes",
            "description": "Top 5 liked notes with scores"
          },
          "top_disliked_notes": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Top Disliked Notes",
            "description": "Top 5 disliked notes with scores"
          },
          "top_accords": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Top Accords",
            "description": "Top 5 preferred accords"
          },
          "top_families": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Top Families",
            "description": "Top 5 preferred fragrance families"
          },
          "llm_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Summary",
            "description": "LLM-generated natural language summary"
          }
        },
        "type": "object",
        "required": [
          "reviewer_id",
          "evaluation_count"
        ],
        "title": "ProfileSummaryResponse",
        "description": "Response model for reviewer preference profile summary."
      },
...
      "RecommendationListResponse": {
        "properties": {
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResponse"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "reviewer_id",
          "recommendations",
          "count"
        ],
        "title": "RecommendationListResponse",
        "description": "Response model for recommendation list."
      },
      "RecommendationResponse": {
        "properties": {
          "fragrance_id": {
            "type": "string",
            "title": "Fragrance Id"
          },
          "fragrance_name": {
            "type": "string",
            "title": "Fragrance Name"
          },
          "fragrance_brand": {
            "type": "string",
            "title": "Fragrance Brand"
          },
          "match_score": {
            "type": "number",
            "title": "Match Score",
            "description": "Match score from 0.0 to 1.0"
          },
          "match_percent": {
            "type": "integer",
            "title": "Match Percent",
            "description": "Match score as percentage 0-100"
          },
          "vetoed": {
            "type": "boolean",
            "title": "Vetoed",
            "description": "Whether fragrance contains a disliked note",
            "default": false
          },
          "veto_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Veto Reason",
            "description": "Reason for veto if applicable"
          }
        },
        "type": "object",
        "required": [
          "fragrance_id",
          "fragrance_name",
          "fragrance_brand",
          "match_score",
          "match_percent"
        ],
        "title": "RecommendationResponse",
        "description": "Response model for a single recommendation."
      },
...
    }
  }
}